renaming and light refactoring to better support adding other harnesses
